+++ src/mod.ts
@@ -4,7 +4,7 @@ export { dev } from "./dev/dev.ts";
 export { fsRoutes } from "./plugins/fs-routes.ts";
 export { tailwind } from "./plugins/tailwind.ts";
 export { staticFiles } from "./server/static-files.ts";
-export type { AppRootOptions } from "./server/ssr.ts";
+export type { AppWrapperOptions } from "./server/ssr.ts";
 export type { RouteConfig } from "./server/router.ts";
 export type { Handlers } from "./server/handlers.ts";
 export type { Context } from "./server/context.ts";